[RT22] Address original RT22 block read errors
While investigating this issue, the original RT22 was often experiencing read
errors while cloning from the radio to CHIRP. It was eventually discovered
that the radio was eventually missing an ACK at some point during the download
which stopped the cloning process with an error.

This patch adds a small delay between the reading a block of data an the
sending of the ACK back to the radio indicating that the block of data was
received OK. The original RT22 radio now receives every ACK reliably.


related to #9043