[RB26] Fix for varying ACK
The RB26 sometimes ACKs the "magic" string with a "\x00\x06" and sometimes
it
ACKs with only a "\x06". CHIRP currently expects to see the "\x00\x06" as
the
response so the process fails if the "\x00" is missing.

This patch "chews up" the "\x00 if present and then only the individual
"\x06" is expected as the ACK.

The Retevis RT76 seems to have the same issue and is fixed as a result of
this
patch.

Fixes #9035