Be defensive about tests that fail to load
The test adapter will fail to load tests if any of them crash on init,
which makes it hard to debug. This adds some output if that happens
and also works around a failure to run get_sub_devices() with no
memobj for those naive drivers that don't do the right thing.