[RT22S] add 6.25 and 12.5 tuning steps
The Radtel T18 and Retevis RT22S radios both support 6.25 and 12.5 steps. The
radtel_t18.py driver does not have a "rf.valid_tuning_steps" entry in
get_features() so the COMMON_TUNING_STEPS list in chirp_common, which does not
include the 6.25 and 12.5 steps, is used by default.

This patch adds an "rf.valid_tuning_steps" entry to get_features() and
specifies the TUNING_STEPS list in chirp_common, which does include the 6.25
and 12.5 steps.

Fixes #9305